moved files to public. commented out db setup to enable page generation. (db config is still wrong
