+++ routes/api-routes.js
@@ -6,7 +6,7 @@
 // =============================================================
 
 // Requiring our Todo model
-var db = require("../models");
+var db = require("./models");
 
 // Routes
 // =============================================================
